\name{CDPriors}
\docType{data}
\alias{CDPriors}

\title{`countData' object derived from data file `simData' with
  estimated priors.}

\description{This `countData' object is derived from the data set
  `simData' and contains the estimated priors. This data set is intended
  to be used to speed the processing of the examples.}

\usage{CDPriors}
\format{A `countData' object.}
\source{Simulation.}

\references{Hardcastle T.J., and Kelly, K. baySeq: Empirical Bayesian
  Methods For Identifying Differential Expression In Sequence Count Data. BMC Bioinformatics (2010)}

\keyword{datasets}
    