\name{densityFunctions}
\alias{densityFunctions}
\alias{ZINBDensity}
\alias{bbDensity}
\alias{bbNCDist}
\alias{mdDensity}
\alias{md2Density}
\alias{md3Density}
\alias{nbinomDensity}
\alias{normDensity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Lists all currently available densityFunctions.
}
\description{
  The \code{\link{densityFunction}} objects define the distribution and
  various other parameters used to analyse the data stored in a
  \code{\link{countData}} object.
}
\usage{
densityFunctions()
}

\value{
  Character string giving names of available
  \code{\link{densityFunction}} objects.
}

\references{Hardcastle T.J., and Kelly, K. baySeq: Empirical Bayesian
  Methods For Identifying Differential Expression In Sequence Count
  Data. BMC Bioinformatics (2010)}
\author{Thomas J. Hardcastle}

\seealso{
  \link{densityFunction}
}

\keyword{utilities}
