\name{factlibsizes}
\docType{data}
\alias{factlibsizes}

\title{Simulated data for testing the baySeq package methods; simulated
  library sizes from a pairwise differential expression analysis}

\description{This data set is a vector of library sizes for the
  \code{\link{factCount}} matrix.
}

\usage{factlibsizes}
\format{A vector containing library sizes for the ten libraries whose
  data is given in the \code{\link{factCount}} matrix.}
\source{Simulation.}
\seealso{\code{\link{factCount}}}

\references{Hardcastle T.J., and Kelly, K (2009). Empirical Bayesian
  methods for differential expression in count data. In submission.
}

\keyword{datasets}
    