\name{mobAnnotation}
\docType{data}
\alias{mobAnnotation}

\title{Annotation data for a set of small RNA loci derived from sequencing
  of grafts of Arabidopsis thaliana intended for differential expression
  analyses.}

\description{This data set is a data.frame (`mobAnnotation') describing
  three thousand small RNA loci identified in a set of Arabidopsis grafting
  experiments.
  
  The data acquired through sequencing for these loci is found in data
  file `mobData'.
}

\usage{mobAnnotation}
\format{A data.frame defining chromosome and position of the sRNA loci.}
\source{Illumina sequencing.}
\seealso{\code{\link{mobData}}}

\references{Molnar A. and Melnyk C.W. et al. Small silencing RNAs in
  plants are mobile and direct epigenetic modification in recipient
  cells. Science (2010)}

\keyword{datasets}
    