\name{simData}
\docType{data}
\alias{simData}

\title{Simulated data for testing the baySeq package methods; simulated
  counts and library sizes from a pairwise differential expression
  analysis.}

\description{This data set is a matrix (`simData') of simulated counts
  from a simple pairwise expression analysis. It is simulated according
  to a negative binomial distribution with varying parameters for each
  row. The first hundred rows of the data are truly differentially
  expressed, the remainder have no differential expression.}

\usage{simData}
\format{A matrix of which each of the ten columns represents a sample,
  and each row some discrete data (acquired by sequencing).}
\source{Simulation.}

\references{Hardcastle T.J., and Kelly, K. baySeq: Empirical Bayesian
  Methods For Identifying Differential Expression In Sequence Count Data. BMC Bioinformatics (2010)}

\keyword{datasets}
    