\name{libsizes}
\docType{data}
\alias{libsizes}

\title{Simulated data for testing the baySeq package methods; simulated
  library sizes from a pairwise differential expression analysis}

\description{This data set is a vector of library sizes for the
  \code{\link{simCount}} matrix.}

\usage{libsizes}
\format{A vector containing library sizes for the ten libraries whose
  data is given in the \code{\link{simCount}} matrix.}
\source{Simulation.}
\seealso{\code{\link{simCount}}}

\references{Hardcastle T.J., and Kelly, K (2010). Identifying Patterns
  of Differential Expression in Count Data. In submission.}


\keyword{datasets}
    